OK-670: Refaktoroitu hakukohde-panel ja henkilo-panel yhteistä koodia erilliseen komponenttiin
