+++ src/app/haku/[oid]/hakukohde/components/hakukohde-list.tsx
@@ -89,5 +89,3 @@ export const HakukohdeList = ({
     </>
   );
 };
-
-export default HakukohdeList;
