+++ src/app/haku/[oid]/hakukohde/components/hakukohde-panel.tsx
@@ -1,104 +1,34 @@
 'use client';
 
-import { Stack, useMediaQuery, useTheme } from '@mui/material';
-import {
-  Close as CloseIcon,
-  KeyboardDoubleArrowRight as KeyboardDoubleArrowRightIcon,
-} from '@mui/icons-material';
-import HakukohdeList from './hakukohde-list';
+import { useMediaQuery, useTheme } from '@mui/material';
 import { useState } from 'react';
-import { notLarge, ophColors, styled } from '@/app/lib/theme';
-import { useTranslations } from '@/app/hooks/useTranslations';
+import { notLarge } from '@/app/lib/theme';
 import { QuerySuspenseBoundary } from '@/app/components/query-suspense-boundary';
 import { FullClientSpinner } from '@/app/components/client-spinner';
-import HakukohdeSearch from './hakukohde-search';
-import { OphButton } from '@opetushallitus/oph-design-system';
-import { DEFAULT_BOX_BORDER } from '@/app/lib/constants';
 import { useParams } from 'next/navigation';
-
-const StyledPanel = styled('aside')({
-  width: '16vw',
-  minWidth: '300px',
-  display: 'block',
-  height: '100vh',
-  borderRight: DEFAULT_BOX_BORDER,
-  top: 0,
-  position: 'sticky',
-  '&.minimized': {
-    minWidth: 0,
-    width: 'auto',
-  },
-});
-
-const ExpandButton = styled(OphButton)(({ theme }) => ({
-  fontWeight: 'normal',
-  height: '100%',
-  color: ophColors.blue2,
-  display: 'flex',
-  flexDirection: 'column',
-  justifyContent: 'flex-start',
-  alignItems: 'center',
-  padding: theme.spacing(1, 0),
-  border: 0,
-  '& .MuiButton-icon': {
-    margin: 0,
-  },
-  '&:hover': {
-    backgroundColor: ophColors.lightBlue2,
-  },
-}));
+import { LeftPanel } from '@/app/components/left-panel';
+import { HakukohdeSearch } from './hakukohde-search';
+import { HakukohdeList } from './hakukohde-list';
 
 export const HakukohdePanel = ({ hakuOid }: { hakuOid: string }) => {
   const theme = useTheme();
-  const isNotLarge = useMediaQuery(notLarge(theme));
+  const isLarge = !useMediaQuery(notLarge(theme));
   const hakukohdeOid = useParams().hakukohde;
-  const [minimized, setMinimized] = useState(
-    () => Boolean(hakukohdeOid) && isNotLarge,
-  );
-  const { t } = useTranslations();
+  const [isOpen, setIsOpen] = useState(() => isLarge || !hakukohdeOid);
 
   return (
-    <StyledPanel className={minimized ? 'minimized' : ''}>
-      {minimized ? (
-        <ExpandButton
-          onClick={() => setMinimized(false)}
-          aria-label={t('haku.suurenna')}
-          startIcon={<KeyboardDoubleArrowRightIcon />}
-        >
-          <span>{t('yleinen.haku')}</span>
-        </ExpandButton>
-      ) : (
-        <QuerySuspenseBoundary suspenseFallback={<FullClientSpinner />}>
-          <Stack
-            sx={{
-              height: '100%',
-              flexShrink: 0,
-              gap: 1,
-              alignItems: 'flex-start',
-              paddingLeft: 2,
-            }}
-          >
-            <OphButton
-              sx={{ alignSelf: 'flex-end' }}
-              onClick={() => setMinimized(true)}
-              aria-label={t('haku.pienenna')}
-              startIcon={<CloseIcon />}
-            />
-
-            <HakukohdeSearch />
-            <HakukohdeList
-              hakuOid={hakuOid}
-              onItemClick={() => {
-                if (isNotLarge) {
-                  setMinimized(true);
-                }
-              }}
-            />
-          </Stack>
-        </QuerySuspenseBoundary>
-      )}
-    </StyledPanel>
+    <LeftPanel isOpen={isOpen} setIsOpen={setIsOpen}>
+      <HakukohdeSearch />
+      <QuerySuspenseBoundary suspenseFallback={<FullClientSpinner />}>
+        <HakukohdeList
+          hakuOid={hakuOid}
+          onItemClick={() => {
+            if (!isLarge) {
+              setIsOpen(true);
+            }
+          }}
+        />
+      </QuerySuspenseBoundary>
+    </LeftPanel>
   );
 };
-
-export default HakukohdePanel;
