+++ src/app/haku/[oid]/henkilo/components/henkilo-list.tsx
@@ -130,5 +130,3 @@ export const HenkiloList = ({
       }
   }
 };
-
-export default HenkiloList;
