+++ src/app/haku/[oid]/henkilo/components/henkilo-search.tsx
@@ -42,5 +42,3 @@ export const HenkiloSearch = () => {
     ></OphFormControl>
   );
 };
-
-export default HenkiloSearch;
